requirements.txt
# Core
torch==2.5.1+cu121
torchvision==0.20.1+cu121

# YOLO
ultralytics

# ONNX & Edge
onnx==1.17.0
onnxruntime-gpu==1.19.2

# Data & Aug
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
albumentations==1.4.0
Pillow==10.4.0

# Visualization & Dashboard
matplotlib==3.9.2
seaborn==0.13.2
supervision==0.24.0
gradio==5.31.0          

# Utils
tqdm==4.66.5
pyyaml==6.0.2
scikit-learn==1.5.2